refactor: ProductStepper로 관심사 분리
