+++ src/recoil/myCartState.ts
@@ -1,4 +1,4 @@
-import { atom, selector } from 'recoil';
+import { atom, selector, selectorFamily } from 'recoil';
 
 import { MyCart } from '../types/Product';
 
@@ -6,17 +6,6 @@ const myCartState = atom<MyCart>({
   key: 'myCartState',
   default: {},
 });
-
-export const makeProductCountSelector = (targetId: number) => {
-  return selector({
-    key: 'productCountSelector',
-    get: ({ get }) => {
-      const product = get(myCartState)[targetId];
-
-      return product ? product.count : 0;
-    },
-  });
-};
 
 export const cartLengthSelector = selector({
   key: 'cartLengthSelector',
